% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amh.R
\name{amh}
\alias{amh}
\title{amh}
\usage{
amh(input, genome)
}
\arguments{
\item{input}{text file derived from output of mhq(yourData) with additional columns as specified below}

\item{genome}{BSgenome object to pull sequences from. IMPORTANT: Has to be loaded before running the function. (i.e. genome <- BSgenome.Mmusculus.UCSC.mm9). Allows any BSgenome supported genome to be used.}
}
\description{
This function is for counting the number of "alternative / more proximal" microhomologies that were bypassed during repair the repair of a CRISPR deletion.
It should be used on Sanger sequenced alleles already analysed by running mhq(input, CRISPResso=F). In the other scenario, When analysing CRISPResso output data, mhq automatically counts alternative microhomologies so this script is not necessary.
}
\details{
\cr
Input is a tab seperated text file (without column names) containing deletions in the same format as output by the mhq function but with some additional columns added. Bed file coordinates are best obtained using a tool like UCSC BLAT.
Also requires the desired genome to be preloaded as a BSGenome object

\itemize{
\item MutantSequence - same as mhq output
\item ReferenceSequence - same as mhq output
\item SizeOfDeletion - same as mhq output
\item NumberOfReads - same as mhq output
\item MH_amount - same as mhq output
\item LD_chr - Additional: chromosome of deletion start
\item LD_start - Additional: bed coordinate start of deletion span
\item LD_stop - Additional: bed coordinate end of deletion span
\item strand - Additional: strand that deletion was mapped to (or strand that microhomology is reported for)
\item sg_start - Additional: bed coordinate start of sgRNAs span (if pairs of sgRNAs used, then start coordinate of 5' most sgRNA)
\item sg_stop - Additional: bed coordinate end of sgRNAs span (if pairs of sgRNAs used, then end coordinate of 3' most sgRNA)
}

\cr
Output is the original dataframe but with two additional columns:

\itemize{
\item altMH_count - Number of alternative microhomologies found
\item add_delSize - Additional size of deletion beyond sgRNAs (same as deletion size when using one sgRNA)
}
}
\examples{
genome <- BSgenome.Mmusculus.UCSC.mm9
amh(input="~/exampleData/Sanger/sequenceDataFile_altMH.txt", genome)

}
